New bill parser approach
Move the bill parsing to the preprocessing, now the output file is divided by segments (as the raw data), however it holds the text snippet and the corresponding mentioned bills. This way we can relate each of the segments to a particular congressmen.

Plus he previous approach was too slow. Now we parallelize the requests so the way raw_preprocessing.py and parse_bills.py interact has changed quite a bit. Still needs refactoring.